+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.7] - 2026-07-19
+
+### Changed
+
+- Add `status`, `seek`, `volume`, and `mute` to the default short field set
+  (`info --fields short`)
+
+
 ## [0.0.6] - 2026-07-16
 
 ### Changed